test: order prof tests by declaration and give pass.go its own test file
Second pass of the suite-wide consolidation. analysis/ was audited and found
clean; no changes there.

§12.1 declaration order — prof/collector_test.go, prof/jit_test.go, and
prof/profiler_test.go now follow their production files. Functions moved
whole; no body edited.

§12.1 file ownership — TestPreserveAll and TestPreserveNone own exported
functions declared in pass/pass.go but lived in pass/manager_test.go. "Each
test file MUST match the production file owning the symbol", so they move to
a new pass/pass_test.go. docs/testing.md:431-432 already recorded pass/pass.go
as their owner, so the matrix now matches reality.

These two tests overlap the "preserve all keeps results" and "drops cached
results" subtests of TestManager_Invalidate, and that overlap is intentional:
PreserveAll and PreserveNone return a struct with only an unexported field, so
their effect is observable only through Invalidate. The two owners assert
different contracts over the same mechanism -- what the Preserved value means
versus how Invalidate honors it -- and both are required by §12.1's one
top-level test per exported function.

Left alone deliberately: prof's Profiler/Collector metric tests, which look
parallel but cover aggregation-after-Flush versus direct storage;
TestProfiler_Flush's concurrent case, per §12.5; and analysis's
TestGVNAnalysis_Run subtests, which assert different fields and would need
per-case assertion functions to tabulate -- the second abstraction layer §12.3
forbids.
